requirements.txt
python-dateutil>=2.8.2
pytz>=2023.3
python-nmap>=0.6.4
dnspython>=2.3.1
cryptography>=40.0.0
pyyaml>=6.0
impacket>=0.10.0
smbprotocol>=1.9.0